Add egde cases
