+++ tests/test_models/test_review.py
@@ -1,6 +1,9 @@
 #!/usr/bin/python3
 import unittest
 from models.review import Review
+import datetime
+from models.base_model import BaseModel
+from models import storage
 """
 Unittest Module for Review class
 """
@@ -39,6 +42,22 @@ class TestUser(unittest.TestCase):
         s = "[{}] ({}) {}".format(self.review.__class__.__name__,
                                   str(self.review.id), self.review.__dict__)
         self.assertEqual(print(s), print(self.review))
+
+    def test_inheritance(self):
+        self.assertTrue(issubclass(Review, BaseModel))
+
+    def test_methods(self):
+        self.assertTrue(hasattr(Review, "save"))
+        self.assertTrue(hasattr(Review, "to_dict"))
+
+    def test_datetime_attributes(self):
+        self.assertIsInstance(self.my_review.created_at, datetime)
+        self.assertIsInstance(self.my_review.updated_at, datetime)
+
+    def test_attribute_existence(self):
+        self.assertTrue(hasattr(self.my_review, "place_id"))
+        self.assertTrue(hasattr(self.my_review, "user_id"))
+        self.assertTrue(hasattr(self.my_review, "text"))
 
 
 if __name__ == '__main__':
